requirements.txt
streamlit
requests
Pillow
mediapipe
opencv-python-headless
numpy